requirements.txt
discord
skinport.py
flask
python-dotenv